requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0